Removed unnecessary synchronisation
Search.solution was unnecessarily synchronising its operations. The only part where it needs synchronisation is the init part. So, we add a (unsynchronized) check there. This will cause the first call in each thread to try to init, but only one thread will get the lock on CoordCube.class and init it. After this init phase, the process will become completely lock-free(actually this is the double-check idiom)